update scaffolding version to v2

+++ util/cache.go
@@ -5,9 +5,9 @@ import "time"
 // 优化 https://github.com/uber-go/guide/blob/master/style.md Avoid Mutable Globals
 
 type CacheTool struct {
-	RandExpire func(int) time.Duration
+    RandExpire func(int64) time.Duration
 }
 
 func NewCacheTool() CacheTool {
-	return CacheTool{RandExpire: RandDuration}
+    return CacheTool{RandExpire: RandDuration}
 }
